+++ src/main/java/my/custom/learning/weblearn/service/PostService.java
@@ -20,7 +20,7 @@ public class PostService {
 		return postRepository.findAll();
 	}
 	
-	public Post findOne(int id) {
+	public Post findOne(Long id) {
 		return postRepository.findById(id).orElseThrow();
 	}
 }
